+++ Day007/Day.cs
@@ -6,13 +6,13 @@ namespace Day007
 {
     public class Day : IDay
     {
-        public double CalculatePart1()
+        public long CalculatePart1()
         {
             var result = 0;
             return result;
         }
 
-        public double CalculatePart2()
+        public long CalculatePart2()
         {
             var result = 0;
             return result;
